requirements.txt
python-telegram-bot
anthropic
requests
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
aiohttp
google-generativeai
